More work to get CI to work.
