+++ Src/Enums.NET.PerfTestConsole/Enums.NET.PerfTestConsole.csproj
@@ -6,7 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="BenchmarkDotNet" Version="0.10.9" />
+    <PackageReference Include="BenchmarkDotNet" Version="0.10.14" />
     <PackageReference Include="System.Diagnostics.FileVersionInfo" Version="4.3.0" />
     <ProjectReference Include="..\Enums.NET.Tests\Enums.NET.Tests.csproj" />
     <ProjectReference Include="..\Enums.NET\Enums.NET.csproj" />
